requirements.txt
# Ingestion
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
tenacity>=8.2.0

# dbt
dbt-postgres>=1.7.0
dbt-redshift>=1.7.0

# Frontend
flask>=3.0.0
folium>=0.16.0

# Dev / Test
pytest>=8.0.0
ruff>=0.3.0
